New GeoJsonGroup layer
