+++ src/layers/GeoJsonGroup.js
@@ -1,0 +1,17 @@
+import { FeatureGroup } from './FeatureGroup';
+import { GeoJson } from './GeoJson';
+
+// Markers layer with labels and buffers
+export const GeoJsonGroup = FeatureGroup.extend({
+    initialize(options) {
+        FeatureGroup.prototype.initialize.call(this, options);
+
+        this.addBuffers();
+        this.addLayer(new GeoJson(options));
+        this.addLabels();
+    },
+});
+
+export default function geoJsonGroup(options) {
+    return new GeoJsonGroup(options);
+}
